Story_Generator_Using_Image/requirements.txt
streamlit
python-dotenv
google-genai
google-generativeai
Pillow
io
BytesIo
tempfile
wave
os
traceback